Change perl version format from -v switch
New format:

  This is perl 5, version 11, subversion 1 (v5.11.1) ...

The rationale for this change is that Perl 5 will never increment to
Perl 6, so we want people to start focusing on the PERL_VERSION number
as significant and as having a series of "releases".  In other words,
"perl 5" is a language, this is the 11th version of it and "subversion 1"
of that version.  (Technically it is the 2nd release, since we're
programmers and start from "release 0", unless we choose to change that
convention.)

The old v-string style is included for reference and usage in code.
For builds from git, it will include the same extended format as it
did before, e.g. "(v5.11.1-176-gaf24cc9*)"