New force_package_version uses scan_version
Instead of using a custom tokenizer to create an old-style dualvar
version, use the version object code's scan_version, and produce
a version object directly.  This is only called by

  package name version;

and requires that scan_version be made more strict about legal
version formats.

Check that all variations of legal and illegal strict version
formats either pass or fail, as appropriate, and with the
expected warning thrown.