Return error strings from Perl_prescan_version
Perl_prescan_version now returns error strings rather than
warning/dying.  It's the responsibility of the caller to
warn or die in an appropriate manner.

This changes the calling signature of Perl_prescan_version
and modifies some macros accordingly.  The change also modifies
the test, as errors arrive in $@ instead of as a captured
warning.