Prevent object methods called as class methods
There are a number of object methods which make absolutely
no sense when called as class methods.  In addition, with
Perl 5.11.5/5.12.0, there are asserts which will trigger
SEGV's when you do that.

So we check in the XS code and refuse to continue if an
object method is called as a class method.